post

+++ app/Http/Controllers/PostController.php
@@ -2,15 +2,29 @@
 
 namespace App\Http\Controllers;
 
+use App\Event\PostCreated;
 use App\Http\Requests\StorePostRequest;
+use App\Jobs\ChangePost;
+use App\Jobs\ProcessPodcast;
+use App\Jobs\UploadBigFile;
 use App\Models\Category;
 use App\Models\Post;
 use App\Models\Tag;
+use App\Notifications\PostCreated as NotificationsPostCreated;
 use Illuminate\Http\Request;
+use Illuminate\Notifications\DatabaseNotification;
+use Illuminate\Support\Facades\Gate;
+use Illuminate\Support\Facades\Notification;
 use Illuminate\Support\Facades\Storage;
 
 class PostController extends Controller
 {
+    public function __construct()
+    {
+        $this->middleware('auth')->except(['index','show']);
+        // $this->middleware('password.confirm')->except('edit');
+    }
+
     public function index()
     {
         $posts=Post::latest()->paginate(3);
@@ -33,7 +47,7 @@ class PostController extends Controller
             $path=$request->file('photo')->storeAs('post-photos',$name);
         }
         $post=Post::create([
-            'user_id'=>1,
+            'user_id'=>auth()->user()->id,
             'category_id'=>$request->category_id,
             'title'=>$request->title,
             'short_content'=>$request->short_content,
@@ -46,6 +60,12 @@ class PostController extends Controller
                 $post->tags()->attach($tag);
             }
         }
+        PostCreated::dispatch($post);
+        ChangePost::dispatch($post)->onQueue('uploading');
+        // auth()->user()->notify(new NotificationsPostCreated($post));
+        Notification::send(auth()->user(),new NotificationsPostCreated($post));
+        // auth()->user()->notify(new NotificationsPostCreated($post));
+
 
         return redirect()->route('posts.index');
     }
@@ -62,11 +82,17 @@ class PostController extends Controller
 
     public function edit(Post $post)
     {
+    // Gate::authorize('update-post', $post);
+    Gate::authorize('update', $post);
+
         return view('posts.edit')->with(['post'=>$post]);
     }
 
     public function update(StorePostRequest $request, Post $post)
     {
+        // Gate::authorize('update-post', $post);
+        Gate::authorize('update', $post);
+        
         if(isset($post->photo)){
             Storage::delete($post->photo);
         }
